# backend/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Distributed Job Scheduler API",
    "version": "1.0.0",
    "description": "API documentation for the Distributed Job Scheduler system"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "slug": {
            "type": "string",
            "description": "Auto-generated from name if omitted"
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "color": {
            "type": "string",
            "description": "Hex color code",
            "example": "#6366f1"
          }
        }
      },
...
      "CreateJobDto": {
        "type": "object",
        "required": [
          "name",
          "queueId",
          "type",
          "payload"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "queueId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "IMMEDIATE",
              "DELAYED",
              "SCHEDULED",
              "BATCH"
            ]
          },
          "payload": {
            "type": "object"
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4,
            "default": 3
          },
          "runAt": {
            "type": "string",
            "format": "date-time"
          },
          "cronExpression": {
            "type": "string"
          },
          "cronTimezone": {
            "type": "string"
          },
          "maxAttempts": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "default": 3
          },
          "jobTimeoutMs": {
            "type": "integer",
            "minimum": 1000,
            "maximum": 3600000
          },
          "idempotencyKey": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "batchId": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/projects": {
      "post": {
        "summary": "Create a new project",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List all projects",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of projects"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "summary": "Get a project by ID",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
    "/api/organizations": {
      "post": {
        "summary": "Create a new organization",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List all organizations",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of organizations"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/organizations/{id}/members": {
      "post": {
        "summary": "Add a member to an organization",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member added"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/queues/{queueId}/jobs": {
      "post": {
        "summary": "Create a new job in a queue",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "queueId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List jobs in a queue",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "queueId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of jobs"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/queues/{queueId}/jobs/stats": {
      "get": {
        "summary": "Get statistics for jobs in a queue",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "queueId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job statistics"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "summary": "Get a job by ID",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/api/jobs/{id}/cancel": {
      "post": {
        "summary": "Cancel a job",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job cancelled"
          },
          "400": {
            "description": "Cannot cancel job"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/api/jobs/{id}/retry": {
      "post": {
        "summary": "Retry a failed job",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job retried"
          },
          "400": {
            "description": "Cannot retry job"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/api/jobs/{id}/logs": {
      "get": {
        "summary": "Get logs for a job",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job logs"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
...
  },
...
}